Refactored Guest entity
