--- domain/src/guests/entities/types.d.ts
@@ -1,6 +1,0 @@
-interface GuestDto {
-  id: string
-  name: string
-  period: DetachedPeriod
-  currently: boolean
-}
