+++ domain/src/guests/useCases/addGuestUseCase.ts
@@ -1,4 +1,4 @@
-import { buildGuest } from '../entities/factory.js'
+import { buildGuest } from '../../models/factory/guest.js'
 import GuestRepository from '../repositories/GuestRepository.js'
 
 export default class AddGuestUseCase {
