+++ domain/src/guests/useCases/listGuestUseCase.ts
@@ -1,4 +1,4 @@
-import GuestEntity from '../entities/GuestEntity.js'
+import Guest from '../../models/Guest.js'
 import GuestRepository from '../repositories/GuestRepository.js'
 
 export default class ListGuestUseCase {
@@ -10,7 +10,7 @@ export default class ListGuestUseCase {
   async execute (): Promise<GuestDto[]> {
     const entities = await this.repository.list()
 
-    const dtos = entities.map((entity: GuestEntity) =>
+    const dtos = entities.map((entity: Guest) =>
       entity.flatten())
     return dtos
   }
