+++ domain/src/models/types.d.ts
@@ -5,3 +5,10 @@ interface ExpenseDto {
   paid: boolean
   period: DetachedPeriod
 }
+
+interface GuestDto {
+  id: string
+  name: string
+  period: DetachedPeriod
+  currently: boolean
+}
