+++ domain/test/utils.js
@@ -1,10 +1,10 @@
-import Expense from '../lib/expenses/entities/ExpensesEntity.js'
-import GuestEntity from '../lib/guests/entities/GuestEntity.js'
+import Expense from '../lib/models/Expense.js'
+import Guest from '../lib/models/Guest.js'
 import PeriodValueObject from '../lib/periods/valueObjects/PeriodValueObject.js'
 
 export const guest = (id, name, from, to, currently) => {
   const period = new PeriodValueObject({from, to, currently})
-  return new GuestEntity({id, name, period})
+  return new Guest({id, name, period})
 }
 
 export const expense = (name, from, to, ammount, currently) => {
